[arch] Refactor: Better separation of concerns and project structure
* http and controllers inside `interfaces` directory
* api service and repository interfaces within domain
* repository implementations and clients within `adapters` directory
* there's no more coupling between domain business definition and
  implementation
* dependency flow goes interfaces -> adapters -> domain
