+++ pkg/interfaces/http/handlers/ping_handler.go
